# src/main/java/com/redia/back/dto/LoginRequestDTO.java
package com.redia.back.dto;

/**
 * DTO para recibir credenciales de inicio de sesión.
 */
public class LoginRequestDTO {

    private String email;
    private String password;

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}